refactor(updater): merge updates/ into updater/ folder
Move RestartDialog from features/updates/ into features/updater/
to consolidate all update-related components in one folder.
Update mock path in home.test.tsx.
